Replace event context with request scope (#23)
This replaces the `event::Context` with a `scope::Context` that includes
extra information about the intent behind the request. The existing
event context data is still preserved but exists behind the
`scope::Context::Event` variant.

The additional scopes are `Admin` and `User`. The `Admin` scope is for
requests coming from the admin interface, providing fewer restrictions
on the accessible data. The `User` scope is for requests from the
login/account selection interface, where data is more freely available
than the `Event` scope but is still restricted to only the user's
access. The `Event` scope is explicitly restricted to the current event.