Extract context from headers independent of axum
Makes the process of extracing the user and scope contexts independent of the axum
FromRequestParts trait. Instead, a TryFrom implementation is added to pull directly from a
HeaderMap. As a result, the FromRequestParts implementation is now simply a wrapper around the
TryFrom implementation.

This also makes the extraction process fully synchronous, increasing the areas in which it can be
used.