Extract context from headers independent of axum (#37)
Makes the process of extracing the user and scope contexts independent
of the [axum `FromRequestParts`
trait](https://docs.rs/axum/0.6.20/axum/extract/trait.FromRequestParts.html).
Instead, a `TryFrom` implementation is added to pull directly from a
[`HeaderMap`](https://docs.rs/http/0.2.11/http/header/struct.HeaderMap.html).
As a result, the `FromRequestParts` implementation is now simply a
wrapper around the TryFrom implementation.

This also makes the extraction process fully synchronous, increasing the
areas in which it can be used.