Use feature powerset for CI matrix
Switch to using a powerset of the crate's features for building to ensure everything works with
everything else.

The matrix elements are generated using a simply Python script that reads the features from the
Cargo manifest. The output is written to a GitHub output as a JSON array, but if the GITHUB_OUTPUT
environment variable is set.