Add current user's role to authenticated context (#46)
Adds a field containing the current user's role to the user context. 

Roles are a general concept that only determines the default permission
level for the user. The consuming service should implement specific
permissions systems. The valid roles are, in descending order of
permissions, as follows: `director`, `manager`, `organizer`, and
`participant`.