Reorganize crate for usability (#47)
Re-organizes the crate to provide better usability and documentation.
All header types are exposed under the `headers` module to reduce
clutter. Similarly, the individual context types are exported from the
root of the crate.

The context types were renamed from `scope::Context` and `user::Context`
to `Scope` and `User`, respectively. This prevents collisions when both
are imported into the same module.

Additional documentation was also added.